Added data in MoreDetail Screen

+++ app/src/main/java/com/example/onlineshop/datamodels/items/MoreDetailsData.kt
@@ -7,7 +7,7 @@ data class MoreDetailsData(
     @SerializedName("description") var description: String? = null,
     @SerializedName("rating") var rating: Double? = null,
     @SerializedName("number_of_reviews") var numberOfReviews: Int? = null,
-    @SerializedName("price") var price: Int? = null,
+    @SerializedName("price") var price: Double? = null,
     @SerializedName("colors") var colors: ArrayList<String> = arrayListOf(),
     @SerializedName("image_urls") var imageUrls: ArrayList<String> = arrayListOf()
 )
